# openapi/openapi.yaml
basePath: /
definitions:
  handlers.LoginBasicRequest:
    properties:
      device_name:
        type: string
      device_platform:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  handlers.RefreshSessionRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  models.APIError:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  models.APIResponse:
    properties:
      data: {}
      error:
        $ref: '#/definitions/models.APIError'
      success:
        type: boolean
    type: object
...
  models.AdminAuthMethodConfig:
    properties:
      bootstrap_required:
        type: boolean
      enabled:
        type: boolean
    type: object
  models.AuthConfigResponse:
    properties:
      admin:
        $ref: '#/definitions/models.AdminAuthMethodConfig'
      oidc:
        $ref: '#/definitions/models.OIDCAuthMethodConfig'
      token:
        $ref: '#/definitions/models.AuthMethodConfig'
    type: object
  models.AuthMethodConfig:
    properties:
      enabled:
        type: boolean
    type: object
...
  models.CurrentUserResponse:
    properties:
      success:
        type: boolean
      user:
        $ref: '#/definitions/models.SessionUser'
    type: object
...
  models.DeviceSessionListResponse:
    properties:
      sessions:
        items:
          $ref: '#/definitions/models.DeviceSessionResponse'
        type: array
    type: object
  models.DeviceSessionResponse:
    properties:
      created_at:
        type: string
      current:
        type: boolean
      device_name:
        type: string
      device_platform:
        type: string
      expires_at:
        type: string
      id:
        type: string
      last_used_at:
        type: string
      revoked_at:
        type: string
    type: object
...
  models.LoginResponse:
    properties:
      access_token_expires_at:
        type: string
      refresh_token:
        type: string
      refresh_token_expires_at:
        type: string
      session:
        $ref: '#/definitions/models.DeviceSessionResponse'
      success:
        type: boolean
      token:
        type: string
      user:
        $ref: '#/definitions/models.SessionUser'
    type: object
  models.OIDCAuthMethodConfig:
    properties:
      enabled:
        type: boolean
      provider:
        type: string
    type: object
...
  models.SessionRevocationResponse:
    properties:
      revoked:
        type: integer
      success:
        type: boolean
    type: object
  models.SessionUser:
    properties:
      auth_method:
        type: string
      email:
        type: string
      name:
        type: string
      username:
        type: string
    type: object
...
info:
  contact: {}
  description: |-
    REST API for managing Garage distributed object storage system
    This API provides endpoints for managing buckets, objects, users, and cluster operations.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Garage UI API
  version: 0.1.0
paths:
...
  /auth/config:
    get:
      description: Returns the current auth configuration (admin and/or OIDC)
      operationId: getAuthConfig
      produces:
      - application/json
      responses:
        "200":
          description: Auth config
          schema:
            $ref: '#/definitions/models.AuthConfigResponse'
        "500":
          description: Configuration unavailable
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Get authentication configuration
      tags:
      - auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate with admin username and password, returns JWT token
      operationId: login
      parameters:
      - description: Login credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginBasicRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            $ref: '#/definitions/models.LoginResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/models.APIResponse'
        "401":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/models.APIResponse'
        "500":
          description: Session creation failed
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Admin auth login
      tags:
      - auth
  /auth/logout:
    post:
      operationId: logoutSession
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SessionRevocationResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - BearerAuth: []
      summary: Log out current device session
      tags:
      - auth
  /auth/me:
    get:
      description: Returns information about the currently authenticated user
      operationId: getCurrentUser
      produces:
      - application/json
      responses:
        "200":
          description: User information
          schema:
            $ref: '#/definitions/models.CurrentUserResponse'
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Atomically rotates a device refresh token. Reusing any previously
        rotated token revokes the device session.
      operationId: refreshSession
      parameters:
      - description: Refresh token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.RefreshSessionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Refresh a device session
      tags:
      - auth
  /auth/sessions:
    delete:
      operationId: revokeAllDeviceSessions
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SessionRevocationResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - BearerAuth: []
      summary: Revoke all device sessions
      tags:
      - auth
    get:
      operationId: listDeviceSessions
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DeviceSessionListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - BearerAuth: []
      summary: List device sessions
      tags:
      - auth
  /auth/sessions/{id}:
    delete:
      operationId: revokeDeviceSession
      parameters:
      - description: Device session ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SessionRevocationResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - BearerAuth: []
      summary: Revoke a device session
      tags:
      - auth
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...